Move contract utils to _utils
