+++ ethpm/_utils/contract.py
@@ -2,20 +2,6 @@ from typing import Any, Dict, Generator, Tuple
 
 from eth_utils import to_dict
 
-from ethpm.exceptions import InsufficientAssetsError
-
-
-def validate_minimal_contract_factory_data(contract_data: Dict[str, str]) -> None:
-    """
-    Validate that contract data in a package contains at least an "abi" and
-    "deployment_bytecode" necessary to generate a deployable contract factory.
-    """
-    if not all(key in contract_data.keys() for key in ("abi", "deployment_bytecode")):
-        raise InsufficientAssetsError(
-            "Minimum required contract data to generate a deployable "
-            "contract factory (abi & deployment_bytecode) not found."
-        )
-
 
 @to_dict
 def generate_contract_factory_kwargs(
